+++ app.py
@@ -7,7 +7,6 @@ window=Tk()
 directory = file.Directory('content/')
 output = file.Directory('docs/')
 extension = '.md'
-files = directory.list_files(extension)
-screen.Screen(window, directory, extension, files)
+screen.Screen(window, directory, extension)
 
 window.mainloop()
