+++ content/git.md
@@ -3,6 +3,10 @@ title: Utiliser Git
 status: publish
 date: 03-03-2021
 ---
+## Tutoriel
+
+[tutoriel](https://realpython.com/python-git-github-intro/) excellent site pour apprendre.
+
 ## Utiliser les branches
 
 Pour travailler en équipe, sans impacter la version principale sur main / master.
