+++ content/jinja.md
@@ -3,12 +3,8 @@ title: Utiliser Jinja2 comme moteur de template
 status: publish
 date: 03-03-2021
 ---
----
-title: Utiliser Jinja2 comme moteur de template
-status: publish
-date: 03-03-2021
----
 ## Sites utiles
+[les bases en anglais](https://medium.com/@jasonrigden/jinja2-templating-engine-tutorial-4bd31fb4aea3) mon préféré.
 
 [site officiel](https://jinja.palletsprojects.com/en/2.11.x/)
