+++ templates/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<!DOCTYPE html>
+
+<html>
+    <head>
+        <meta charset="utf-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <title></title>
+        <meta name="description" content="">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <link rel="stylesheet" href="">
+    </head>
+    <body>
+        
+        <!-- ajouter une image en fond (<span>Photo by <a href="https://unsplash.com/@sigmund?utm_source=unsplash&amp;utm_medium=referral&amp;utm_content=creditCopyText">Sigmund</a> on <a href="https://unsplash.com/s/photos/pen-and-paper?utm_source=unsplash&amp;utm_medium=referral&amp;utm_content=creditCopyText">Unsplash</a></span>)-->
+        <!-- ajouter texte de ReadMe -->
+
+        <ul>
+            {% for title in titles %}
+                <li><a href=" {{ title.url }} "> {{ title.label }} </a></li>
+            {% endfor %}
+        </ul>
+        
+        <script src="" async defer></script>
+    </body>
+</html>
