+++ templates/page.html
@@ -9,7 +9,8 @@
         <link rel="stylesheet" href="antiseche.css">
     </head>
     <body>
-        <nav></nav>
+
+        {% include 'navigation.html' %}
         
         <main>
             {{ content }}
